test(bootstrap): consolidate tokenless auth coverage
